requirements.txt
streamlit
pycryptodome
Pillow
matplotlib
numpy
